[PVAD2GRFFListModel] one line bug-fix (and 12 lines comment:)
the following line is REQUIRED to avoid a segfault when doing the
following step:
1- click on "Add"
2- click on "Up" or "Down"
In that case, remove_Rows() frees the a RFF (stored as a shared
pointer) because it is its last reference. setData()
(shared_from_this more precisely) throws also the following
exception:
"""
terminate called after throwing an instance of 'boost::exception_detail::clone_impl<boost::exception_detail::error_info_injector<boost::bad_weak_ptr> >'
"""

In the other cases, the PVAD2GFunctionPropertiesWidget is shown,
it has also a reference on it, so that the RFF is not freed.