[PVHive] work-around for Qt's hackers thinness
Qt's signals/slots mechanism can not work properly with namespaces;
leading to run-time errors of type "Incompatible sender/receiver
arguments" or "No such signal": the signals use implicit namespaces
prefix (otherwise it does not compile) and the slots use explicit
namespaces prefix. So that the MOC's internal strcmp fails when
comparing signals/slots signatures.

Adding "using namespace NAMESPACE" solves the problem but it is error
prone.

Adding the new derived class in the same namespace is not acceptable.

So here is a hack to have PVRefreshSignal defined in its namespace *and*
the signals defined outside any namespace: use a (very) long and unusual
class name for the class definition and use "typedef" in the namespace.

The new signals/slots mechanism of Qt 5 may solve this issue.