[PVHive] usage is more generic
When dealing with Qt signals in a multithreaded program, each thread
calling "emit" must inherit from QObject and use an event loop (A
Qthread). Signals can also migrate from a thread to another; otherwise
the program may crash (under Linux, XCB complains about XinitThreads()
not called in a multithreaded program).

Rewriting all threads of Inspector as QThread can be a mess. So that the
PVHive now inherits from QThread and the behaviour of call_object() and
refresh_observers() have changed.

These functions are now fully asynchronous: the actor now asks the hive
to run the action for it.  Since an action can affect a Qt object, it
may leads to a crash as describe above.

So that, an actor can be in a boost::thread, an OpenMP one or a Qt one.