[PVHive] reinforcement of ::unregister_object(T const &)
As a template method, it can accept any type. If T is a pointer, the
behavior is erronous: the address of the pointer is used instead of its
value.

Add a static assert to check that T is not a pointer.

thx to JB for the tip (std::enable_if was not the best way to go).