[PVHive] neat synchronous signal emission
Qt's signals are emitted asynchronously when the emitter and the
receiver are not in the same thread. To make sure the call is
synchronous, a first signal is emitted to "move" to the right thread and
a lock is done on a semaphore. In turn, the intermediate signal emits
the effective one and releases the lock.

If emitter and receiver are in the same thread, the effective signal is
emitted directly because it is necessarily synchronous.