better user experience when scrolling the zoomed view
When the scrollbar is used to pan the view, the update can be fast
enough to have a flickering visual effect. Using a timer to delay the
full update helps removing this disturbing effect.

In other cases of view panning, the timer give the impression of slow
refresh. So that it is not used in those cases.