Faster rendering for zoomed view
When searching for selected lines in the zoomed view, the ZoneTree
informations can be used in the ZoomedZoneTree to avoid traversing
quadtrees for nothing : ZoneTree._sel_elts contains the index of the
first selected line for each bucket (or quadtree). Its value can also be
used to know if a traversal has to be done (_sel_elts contains a valid
index) or not (an invalid index).

It implies 2 important changes:
1/ PVLibView::process_selection_Observer::update() must be used to
   tell PVZoomedParallelScene thats their selection images are invalid
   instead of telling PVZoomedParallelScene it has to to redraw itself

2/ a ZoomedZoneTree can be processed only if its related ZoneTree has
   been processed. PVZoomedParallelScene is also connected to the Qt
   signal PVZonesManager::filter_by_sel_finished

others minors changes that had to be done:
* PVZoomedParallelScene::zone_desc -> PVZoomedParallelScene::zone_desc_t
* zone_desc_t.image -> zone_desc_t.bg_image (coding line)
* PVZoomedParallelScene::zone_desc_t has a field "created" to avoid
  testing if zone_desc_t.bg_image.get() is null or not
* adding some comments for PVZoomedParallelScene's member variables