PVZoomedParallelScene limits its view height to 1024
A note is in PVZoomedParallelScene.cpp. But here it is:

when zooming, the smallest backend_image's height is 1024 (2048 / 2). So
with a view's greater than 1024, there will be an empty space at the
view's bottom. So the view's height must be limited to 1024. And
limiting its width to 1024 is not so weird. The QGraphicsPixmapItems
height has to be limited to 1024 too: the restriction for BCI codes is
only on the zoomed axis, not on their neighbours. The
QGraphicsPixmapItems could also contains a trapezoid containing lines,
which is ugly. Having a height of 1024 remove this problem.