[PVHive] partial bug-fix when observers access the hive while refresh
The hive's goal is to add observers in an object modify it. But when
calling PVHive::call<...>(...), the observer may add new observers.
In that case, a deadlock occurs because the iteration has a lock
and the call to PVHive::register_xxx try to acquire another lock (and in
the same thread) and the mutex used in tbb::concurrent_hash_map is not
recursive.