[all] add combined axes identification
In PVAxesCombination, a combined axes could only be retrieved with its
index. A zoom parallel view can also not retrieve its axis it it has
been
moved because more information is needed.

A combined axe has now an id (a union of its original axe and a counter)
to help retrieving the index of a given id.

PVCol has now an invalid value (-1)

The type used as combined axe's id is theorically compatible with old
format when (de)serializing.