[PVParallelView] better lines extraction in zoomed view
The current algorithm extracts a set of relevant events before
converting them into BCI codes, finally a collision test is done.

Theorically, the number of BCI codes is lesser that 6 Mi but in
practise, more lines are send to the CUDA kernel (when a segfault did
not occur). The reason is that the collision test was based on a too
weak assumption about events order.

The new algorithm does not use a cache to store extracted events and a
BCI codes buffer. It uses a buffer which only stores events index and
color; the rest of the BCI code (type, l, and r) is implicitly encoded
into the index used to access buffer's value. Datas are also not
contiguous in this buffer. So that, for each relevant events, its
type/l/r are compute and use to collide it into the partial BCI codes
buffer. After the forest traversal, the buffer it traversed to create
the final BCI codes.

minors things:
* some parameters/types/whatever renaming
* quadtree traversal functions do not return anymore the extracted
  elements number
* updated test program (to make it compile without error)