[PVParallelView] PV{Full,Zoomed}ParallelView have now a minimum size
PVFullParallelView's initial size is too small to be usable. Using
QSizePolicy does not help changing this behaviour. This problem is
reported in many forums but there is no solution.

So that, the PVFullParallelView has now a minimium size to constraint
QMainWindow to let it enough place on the screen.

To be clear, PVZoomedParallelView has a minimum size too.