[PVRush] bug-fix: PVNrawDiskBackend::flush was called twice
It leads to a double free when using Source/Extractor on an already
loaded file.

PVControllerJob::job_has_run() was also called twice on the same object
(but it must never happen). In case of PVControllerJob::read_everything
in PVController::operator(), job_has_finished() must not be called on
the related PVOutput (which calls PVNrawDiskBackend::flush).

and:

* added some "ptr = nullptr" after deallocation in PVNrawDiskBackend.cpp
* remove useless code in test program rush/process_file.