[PVParallelView] simplify zones zooms in full views
Using global and per zone zoom level can be annoying because the wheel
mouse may have no visual effect due to the clamping of zones width.

So that, there is only one zoom level: the per zone one. The global zoom
increase or decrease (respect to the accepted range) the zoom level of
each zone. Users won't be disturbed by the "ineffective" wheel mouse
action.

The reasons of this simplification are:
* this anoying visual effect
* the preservation of zones aspect-ratio is a good idea must it must be
  think back from scratch