[PVParallelView] layout zones according to the full views viewport
When zones can fit in the viewport, they are resize to be in a
undersantable way for the user (centered and resized to fill the
viewport); ortherwise the default width is used and zones are aligned to
the left.

Use of "Home" key to reset the display is back.

minors changes:
* a typo in a comment
* FullParallelScene::keyEvent() did not honor event acceptance/passing