[PVFilter] close ticket #193 "Détection d'un splitter CSV très lente dans le format builder"
The slowness had 2 origins:
(1) the algorithm to detect the best splitter configuration had a
    complexity in O(n²)
(2) all the content of a PVChunk is tested, it can also result into a
    big set of potential configurations to test.

This commit solve the problem (1).

The goal of the rewritten code is to find the pair (argument list, field
number) among the possible configurations deduced from (2) whose
argument list is unique in the configuration set (having 2 different
field numbers for a given argument list indicate the extraction will not
be stable).

The new algorithm use a different approach: counting the occurrence
number for each pair (argument list, field number) by inserting them in
a double hash-map. removing all pairs leading to an unstable extraction
and retrieve the one with the highest occurrence number.

An example can be found in PVFilter::PVFieldSplitterChunkMatch::get_match()

some statistics printing are added too (BENCH_{START,END} and dump of
the reduced candidates set.