[PVParalllelView] close ticket #219 "the selection rectangle reappears when resetting the zones width"
In fact, the selection rectangle reappears each time at least one zone
is resized (i.e. when the zones are "zoomed").

The selection rectangle is also shown when needed and hidden when
unneeded (shown and hidden in QGraphicsItem's meaning). When shown, its
geometry is reset to correct value (the current mouse position, not the
rectangle's previous position).