[PVParallelView] close ticket #200 "Columns width and zones width are wrong after changing axes combination"
After an axes combination changing, the zones widths are not updated:
they do not follow the zones when they are moved. It implies all the
annoying things which can be observ.

When an axes combination have been changed, the PVFullParallelScene
(through the method ::update_number_of_zones()) reorder the graphical
axes to make them be at the right place (and delete unneeded axes or
create new ones).

The zones width preservation follows the same process (but it is about
zones, not axes) : resynchronizing elements states.

So PVFullParallelScene::update_number_of_zones() has been update to
preserve zones widths.