[PVGui-Qt] faster PVSelection computation from PVListingView
The old method extracts the list of selected row and iterates on it to
set the selection. Due to the model/view aspect of QTableView, the
extraction takes time and memory.

The new approach omits the list extraction and traverses directly the
entry set to create the selection. Finally this algorithm is
parallelized with a parallel reduce.