[all] better mouse tracking support in PVGraphicsView
QGraphics has a strange working: the view's painted widget is the
viewport, so that the mouse-tracking property must be set on the
viewport.

But for Qt's end-users (us), doing a paint event on another widget than
'this' is stricly forbidden. To have a working mouse-tracking behaviour,
PVGraphicsView must set this property on the viewport *and* the view.

So, PVZoomedParallelScene do not need to set this property on its view.