[PVWidgets] bug-fix for the PVGraphicsView's phantom translation
When a resize event occurs, the scene update itself, which causes
indirectly a call to PVZoomedParallelScene::update_zoom() which updates
the transformation matrix; combined to the use of AnchorUnderMouse as
the transformation anchor, the view was centered on a point outside of
the view. The forced horizontal centering give the impression of a
vertical translation related to the mouse movement.