[PVParallelView] bug-fix in PVGraphicsView
PVGraphicsView::event(...) automatically calls for an update except
for a set of specific event types. At this point, QEvent::Paint seemed
to be the lone.

But the widget was in an unwanted "auto-refresh" mode. That was due to
the fact that PVGraphicsView::event(...) did an update for
QEvent::UpdateRequest events.

So, this event does not imply umpdate anymore.