close ticket #243 "Removing temporary nraw when starting the application"
When PVI crashes, the nraw files are not deleted as usual. To avoid
filling the nraw temporary dir, unused directory are removed each time a
new instance of PVI is started.

As PVI keeps its nraw files opened as long an investigation is used,
unused directories can be defined as : directory containing files
which are not opened by a running instance of PVI".

Unices standard authorizes that a file can be opened by more than one
program. To detect such case, Linux allows to lock a file to prevent any
other programs to access it. As a lock can not be set on a file if it
has already been opened by an other programs, this help detecting unused
files (and directory).

A new file helper has been added with this test : PVCore::PVfileHelper

PVRush::PVNraw has been enhanced with method to remove its unused
directories.

The removal is done while the splash screen is visible.

2 tests have been added to validate PVCore::PVfileHelper's method and
the PVRush::PVNraw::*_unused_nraw_directories logic.