close ticket #246 "Crashes with invalid format file"
As the invalid format problem will be solved with the format builder v2,
the hack used in the CSV splitter is reused in the Duplicate and RegExp
splitters: if the extracted field number does not match the expected
number, the field is reported as invalid; discarding all events. The
resulting effect is that there is no valid extracted events.