[PVZoomableDrawingArea] axis invertion support
* add support for y or x axis invertion
* invert 'x' axis of the scatter view, so that it is coherent with other
views. Please note that the plotted is inverted so that the full and
zoomed parallelview have their '0' values at the bottom. This is just a
hack over the original hack, and the cleaner thing to do is to render
the BCI codes within the good space coordinate. With that, you also need
ot invert the hit count y axis.