[PVgraphicsView] bug fix
Give the associated QGraphicsScene a default view transformation so that
QGraphicsScene::itemsAtPosition can take into account "untransformable"
QGraphicsItem (that is ones that mustn't be affaceted by the view
transformation).
This is a hack as only one view can be used with one QGraphicsScene. The
thing is that QGraphicsScene is tied to QGraphicsView. The real fix
would be to add a "proxy class" that just has the virtual methods that
QGraphicsScene needs, and implement them in QGraphicsView and
PVGraphicsView.