fix bug added in commit 0754382f9fe3b0e7b532a27127c69df70561d192
whom title is "[PVGuiQt] the layer stack displays the number event of the
layers"

I did not tried to open an investigation (which is way to create layers).

The bug is due to the fact that, in case of .pvi load, the PVView (and its
layer-stack) are created before the mapped is created; its size is still 0
and counting bit in a PVSelection does not like using 0 as a size parameter.

So, the fix is to add a finalization step after the load of a mapped (to
finish things in PVPlotted, and in PVView which update its layer-stack)