[PVWidgets] add PVHelpWidget
Help widgets was initially PVTextPopupWidget but the key needed to open it (or
to close it) was hard-coded in PVTextPopupWidget. A realy bad idea. Now,
PVPopupWidget provide a virtual method is_close_key() to override like in
PVHelpWidget because PVTextPopupWidget's ::eventFilter() need a virtual
method to work... I think I'm not clear O:-)