[PVGuiQt] less space used by input plugins in start screen
In the start screen, input plugins were displayed in a vertical layout.

With an horizontal layout, it is far better but not the best: A grid should
be far better with a fixed column number and an ::appendItem() to fill
automatically the layout. Defining our own order would be useful too.