[PVParallelView] bug-fix in PVZoomedZoneTree::browse_trees_bci_by_y1_y2_tbb(...)
In some case, the insert_entry_y1_y2_f(...) lambda function is called with an
invalid event (its coordinates are irrelevant) and raise an exception.

The fast wat to solve this problem is to reject those bad events but this
bug's origin has to be found and solved.

some perf code has been moved in #if PICVIZ_DEVELOPER_MODE blocks