[script] extend max opened file count to a higher value
In some cases, PV-I reports "too many opened file". This can be explain by
the plugins (one file descriptor per plugin) and the Nraw (many file
descriptors per column and/or per thread).

A work-around is increasing the user's max opened file count by using the
shell's builtin "ulimit". It is set to the highest possible value.

The next step is to a file in /etc/security/ulimits.d to permit higher value
for the picviz user.