[Gui-Qt] splitting preview now always does a full reload
Clicking on the Preview button in the format builder may crash PVI (as it is
a widget instead of an external program). The crash origin seems to be related
to synchronization problem between the format representation and the Nraw
structure.

The fix is simply to avoid updating the nraw part and redo a full reload from
the source.

It does no fix load crash.