[PVWidgets] "better layout" for mapping/plotting editor
The mapping/plotting combobox's width is too big compared to the others ones.
In fact, this width is what it is because a string mapping has a long name:
"Pseudo-sort on the first 4 bytes". To have an homogeneous layout, all the
the mapping/plotting comboboxes must have the same width...

The different involved layouts borders/spacing were not elegant. It's solved.