[PVFieldSplitterDnsFqdn] close ticket #274 "Splitter : DNS hostname splitter"
It permit to split a FQDN in a DNS context into:
- TLD1: the first top level domain
- TLD2: the two first top level domains
- TLD3: the three first top level domains
- SUBD1: all but the first top level domain
- SUBD2: all but the two first top level domains
- SUBD3: all but the three first top level domains

In case of PTR DNS query, the third level domain is an IP address (the
SLD+TLD is also "in-addr.arpa" for IPv4 and "ip6.arpa" for IPv6); the
fields behavior is also:
- TLD3 remains empty
- IP addresses appear in the SUBD* fields

When SUBD* contain IPs, they can be reversed to look in the "right order".

Note: the involved TLD (top level domain) concept has be understand in its
original form: the token after the last dot ('.'). More recent TLD concept
like "co.uk" or "notaires.fr" are not handled as special cases.