Close ticket #274 "Splitter : DNS hostname splitter"
It permit to split a FQDN in a DNS context into:
- TLD1: 'com' in 'www.en.example.com'
- TLD2: 'example.com' in 'www.en.example.com'
- TLD3: 'en.example.com' in 'www.en.example.com'
- SUBD1: 'www.en.example' in 'www.en.example.com'
- SUBD2: 'www.en' in 'www.en.example.com'
- SUBD3: 'www' in 'www.en.example.com'

In case of PTR DNS query, the third level domain is an IP address (the
SLD+TLD is also "in-addr.arpa" for IPv4 and "ip6.arpa" for IPv6); the
fields behavior is also:
- TLD3 remains empty
- IP addresses appear in the SUBD* fields

When SUBD* contain IPs, they can be reversed to look in the "right order".

Note: A FQDN may start with a dot ('.') but it is insightful. As it can
have a meaning for the user, it is not removed.

Note 2: A FQDN may end with a dot ('.') but it is insightful. While it has
a meaning while processing a hostname (to expand it with the domain name
or not), it has none in any other cases. It is also removed from the
splitting proccess (and result).

Note 3: TLDs like "co.uk" or "notaires.fr" are not handled by this
splitter.