[general] Make integers axes correctly sortable in listing view
A string sorting algorithm were used anytime; but there is a plugin class
for sorting functions (according to axis' type and mapping (I guess)).

The listing view now requests the good sorting function (with a fall-back
to the string one) to do the sorting.

Need some extra sorting plugins (for hexadecimal, octal or ipv4).

It's is f***ing slow...