[PVKernel] close ticket #298 (Fix crash at 200 000 000 lines)
The import pipeline loads chunks as long as the source's EOF is not reached
(since PV-I 3.2, not tested before). The mapping/plotting/views/... use the
PICVIZ_LINES_MAX constant but not the statistics views (SV) which use the
Nraw's internal count. So that, starting a distinct value SV will iterate over
the NRaw (using its internal count) using a selection (which is correctly
limited)... segfault!

The ouput stage of the import pipeline now write at most PICVIZ_LINES_MAX
events in the NRaw *and* tell the input stage it can stop loading.

As there is/are still chunk(s) to process, the import does not stop immediatly
(as when the import is interrupted by the user to load partially the source).