[PVCore] extend QObject helpers list
to do screenshots of views, it is useful to search for a relevant
widget to capture: the QDocWidget around the listing. But the helper
named PVCore::get_qobject_parent_of_type() is not suitable because it
starts searching with the parameter's parent, not its parameter.

PVCore::get_qobject_hierarchy_of_type() do the wanted job (and
PVCore::get_qobject_parent_of_type() uses it).