[PVGuiQt] close ticket #205 (Split unique values into layers.)
This functionality automates the following actions:
- do a new selection with a search
- create a new layer using the selection

This request has been generalized to all of statistic views and implements 2
possible actions:
- create a new layer with the result of the search for one or more values
- create as many new layers as there are selected values

The naming has to permit to use substitution pattern according to the
environment (layer name, axis name, selected value(s)).

The placement behaviour of the new layer(s) can be parametrized too:
- on the layer-stack top
- just above the current layer
- juste below the current layer.

About the implementation, it is trickier...

As the search is done with a plug-in which is dynamically append to the context
menu, a hack has been used to help finding the wanted QAction to launch the
right multiple-search code.

As the layer-stack exposes its index based logic and is too centered on the
selected layer, the layer placement logic is outside of it (which is not
really pleasant) but there are too many changes to do on the layer-stack to
do it now.

As PVView is the last class usable with the hive, it is used (as usual) as
a god objet to access the layer-stack.

The many uses of the Hive are required to have a working functionality.