[converter_value_mapper] Fully working implementation of the Value Mapper Converter.
The following parameters are available:
- Specify a conversion file. A warning is issued if unable to open the file, or no value mapping are parsed.
- Specify a separator char and a quotation char.
- Specify a default value for values without mappings.

The CSV file is parsed with libcsv and the conversion to UTF-16 is done once using ICU and stored into a std::unordered_map.
Note: libcsv is duplicated with 'splitter_csv'. Need to find a proper place to put this lib...

In order to do some needed preprocessing, a virtual init() method was added to PVFilter::PVFieldsFilter and called in PVRush::PVFormat::xmldata_to_filter.