[PVRush] close ticket #294 (CSV splitter: ignore first lines)
This patch has been mainly done by JBL

A new option has been added in the format files to tell how many lines have
to be skipped when importing from a PVInput. In case of aggregations, the
line skip is done for each PVInput, not just for the first one.

An option has been added too to tell how many events have to be imported but
due to a bug, this limit is not properly used. This will be solved in an
other patch.

Some minor indentation changes too.