[PVParallelView] ZPV's dock title can not be updated when the axis changes
When the axis is changed in the ZPV, the title of the parent dock widget
should be updated.

But Qt < 5.2 does not have signal/slot for window's title/icon updates. Just
an event which has to be caught and propagated through many classes... which
sux...

The current's axis name is display as an overlay (like in FPV), to nicely
solve when we move to Qt 5.2).