[Gui-Qt] fix a segfault when doing a screenshot
Doing a screenshot of the start screen leads to a nice segfault...

The reason is that the parent widget is PVGuiQt::PVWorkspacesTabWidgetBase,
but the PVProjectsTabWidget does not have any current workspace to get a
name to use for the screenshot filename...

As we are either on a data collection (which has at least one source) or
on the workspaces (which always have a default workspace), we can have a
string to create the screenshot's filename. If we are on the start screen
(no "tabwidget"), "startscreen" is used in the screenshot's filename.