[Picviz] improved IPv4/uinform mapper
before, a hashmap was filled with encountered IP addresses and in a
finalization pass, the nraw was traversed to use the hashmap in
mapping computation.

Now, the mapping storage used in the import is used to store the IP
as a numerical value. The finalization pass iterate on it (instead of
the nraw). This avoid converting the IP from string twice and avoid
having "imaginary" IP in the hashmap.

It removes the bug of holes in the mapping when the import is interrupted
(with the max line count or by pressing the "stop and process" button).