[PVGuiQt] fix i-BP's bug about empty selection when doing a "search using this values"
step to see the bug:
1/ open a stat view
2/ pick a range which produce a non empty selection
3/ do a "search for this"

The selected events set is necessarily empty.

The regression has been introduced in the commit 27a930046cec451778b641494c6b5109a1aca844
which replaces a call to
	selection_model->selection().indexes();
with a call to
	selection_model->selectedRows(0);

This seemed to be equivalent (and a little bit easier than using the
selection) but it fact it is not: in case of selection done with the mouse or
the keyboard, the two methods return the same result but if the selection is
generated from the range picker, they have different results...

Reverting the change solves the problem.