[PVQuiQt] fix failing search in stat views with empty strings
When there is more than than one value in the search, if one of them is the
empty string (because there is no value in the nraw), the regexp to match
empty value was not set properly.

The multiple_search plugin now takes care of multilines expression with "empty
string".

This bug-fix permits to see that "" were used in layers names, "(empty)" is used
instead.