[PVParallelView] close ticket #313 (Zoomed parallel coordinates views crash when changing too fast the displayed axis)
Cancelling a zone rendering leaves the rendering pipeline in an inconsistent
state because the end of a task is not really effective when the cancellation
returns. So that, a rendering can be processed even though it has been
cancelled...

In that case, the used PVBCIBackendImage are freed while the rendering still
needs them... not fair...

Using shared pointer on PVBCIBackendImage in the whole rendering process
instead of simple pointer or reference solves this bug.