Fixed a crash in format builder.
When loading a sample file for preview that has a key/value splitter,
the filter init() method wasn't called, causing a segfault.

This was due to the fact that the format builder has it own code path
for parsing the data. This is fondamentally wrong and will be fixed for
the next iteration of the format build when integrated with libpvcop.

https://gitlab.srv.picviz/picviz/picviz-inspector/issues/1