Make the correlation system optional at build time
As this functionality depends on an old Tulip version which is Qt 4 based,
porting Inspector to Qt 5 implies upgrading the Tulip whom API has totally
changed (we can consider we would have to rewrite an huge part of this
functionality).

An other point is that the correlation is still experimental (slow, not well
featured) and will need more work for pvcop and typed axis integration.

As this feature is not really used by customers, we choose to disable this
functionality. It stays in the source code but it is disabled by default at
build time. Running cmake with -DENABLE_CORRELATION gets it back.