[Qui-Qt] Simplify file+specific format load code path
Since the correlation is optional, picking a file *and* a format file
leads to a segfault due to a null current PVScene (when there no one created).
Loading a source description did not crash because there were 2 code paths:
* PVMainWindow::load_source_from_description_Slot(...) for a source
  description
* PVMainWindow::import_type(<long parameter list>) to do the full processing
  (autodetection, local formats use, specific format)

As those 2 methods do PVScene retrieval, PVScene registration in the
data-tree, and lots of similar things until the call to
PVMainWindow::load_source(...) but not the same way, the 2 code path
could be factorized to only use ::load_source_from_description_Slot(...).
This method merges both error handling and ::import_type(...) now calls it.

It also permits to select the target data collection to load the new data set
to.