[FIX inport_type] Do not crash when importing file with custom format
* Hash changed with Qt5 so order in QHash is not predictable. While
"text" was the last one while iterating on this hash map in file format
mode, it is not the case anymore. For now, we force to use "text" in
these cases but it disable easy use of pcap, perl and Python. This
modification show that there is a design issue in the code and we don't
really know what have to be done.