fix crash in REST plugins
Inspector crashes on Wheezy when exporting data through the Elasticsearch
connector ; same problem occurs with the Splunk one.

It happens in PVParamsWidget:: export_slot(...) when it calls the virtual
method PVParamsWidget::export_query_result(...). the class instance (this) is
also reported as null... It seems that mixing template classes, virtual
methods and lambda functions is a little bit complicated :-]

As this bug does not occur with GCC 4.9 (under Jessie), The lone reason
is that GCC 4.7.2 (under Wheezy) has a bug. But the truth can be out there...

Using an explicit `this->method(...)` solves the bug.

one line patch! \o>