[QueryBuilder] fix PVQueryBuilder which can use old values
If you change a value in the query graph, starting a request will not use
this value but the previous one. To prevent that, you have to change the
focus elsewhere.

This bug happens because JQuery-QueryBuilder seems to use the *change* event
to update its internal structure.

So that, forcing a *change* event on the active DOM element permits
to have a up-to-date query.

close #46