[InputTypes] local file plugin may crash when opening its file dialog
Under proto-03, Inspector (from master or branch-3.5) crashs when opening the
file dialog of the local file input type plugin... It's strange because it
does not happen under remote access to proto-01b and keynux (respectively
under Wheezy and Jessie). It can be a problem with Gtk+ (I use XFCE as
desktop manager).

The Qt documentation explain that in our case, the native file dialog is not
used but it is what happens...

Forcing the option QFileDialog::DontUseNativeDialog solves the problem.