[PVRush] make import stop at the right max events count
As exlained by @jteliet, the last chunk was fully written to the disk when it
had to be partially written according to the remaining events count.

The conversion loop from PVCore::PVElement to a vector of pvcop::field_t did
not do the job as it was based on the PVCore::PVChunk geometry instead of
its content.

close #116