[PVKernel] fix investigation opening from crashing
As PVCop is now used, the old Nraw structure does not match PVCop's one;
leading to an exception (no description file found).

As PVCop now throws a specific exception when its root directory is
dadly-formed, we can catch it to mimic the absence of a nraw directory and
force its regeneration in the right format (like before when the nraw
directory were removed to save space).