[PVRush] reduce import memory consumption
PVExtractor was initialized to used 5 times the number of cores for the number
of live tokens in a TBB pipeline. Using only the number of cores as tokens
count does not change the import time but it reduces the memory consumption.

On proto-03 with the proxy_sample.log file (10 Me), the max memory usage is
(for a given multiplier):
- 5: at most 11.7 Gio are used;
- 2: at most 7.2 Gio are used;
- 1: at most 5.3 Gio are used.

A file with 2 columns of 0 makes swap proto-03 at 65 Me (63 Gio used).