[PVGuiQt] fix bug when copying to clipboard in PVStatsModel
The copied values are correct but not their related statistics.

The bug has been introduced in commit d0c0d2fd1f9dd2b3257067be21920e1ba678b7ad
which does not use the same index to read data from the two pvcop::db::array.

The problem comes from the fact that the method PVStatsModel::export_line(...)
has to be used with indices in the PVAbstractTableModel internal selection
and not with indices of cells displayed in the QAbstractItemView. The use of
the method PVAbstractTableModel::row_pos_to_index(...) was also erroneous.