Removed PVSelBitField constants related to INENDI_LINES_MAX
This implied modifying PVSelBitField and PVLayer constructors to take
the number of rows as parameters.

As sometimes default constructors needs to be called, the PVSelection
row count is set to INENDI_LINES_MAX if not specified otherwise.
The proper value should then be set using PVSelBitField::set_count
before any memory allocation occurs (as PVSelBitField::allocate_table
is called lazily).
If somehow, the allocation is done anyway using INENDI_LINES_MAX row count,
an assert will pop in debug mode to fix the problem; that means everything
should continue working as usual but with smaller allocation size.

The ultimate goal is to get rid of default constructed objects but it
can't be achieved in one step...