[format-Builder] fix repeating popups bug
The first time a format-builder is created, it works correctly. The second
time a format-builder is created, plugins QActions are triggered twice. If N
format-builders have been created, the QActions will be triggered N times...

The bug origin is that the same QAction can be considered as a static member
variable of the plugins. the different created format-builders are also
connected to the same pair QObject/signal. Triggering a QAction every
connected slots. An other potential problem is that the call to old slots can
lead to a segmentation fault.

Correcting this bug implies:
- plugins create and return a new QAction everytime ::get_action_menu(...) is
  called instead of always returning the same one;
- plugins create QAction whose parent is a format-builder QWidget, not a
  plugin one;
- PVXmlTreeNodeDom::initSplitters(...) and PVXmlTreeNodeDom::initMenuBar() are
  merged to centralize the plugins list processing.

This patch includes too:
- some cleanups in PVXmlTreeNodeDom::initMenuBar();
- adds some "override" in the plugins classes to make thing cleaner;
- removes a double QMessage popup.