[URL-Splitter] fix crash when omitting child axes
This splitter had not been ported to axes omission... my fault...

This splitter worked because it was in the good circumstances to do it job:
the fields count must be equal to the tagged axes count.

Axis omission breaks the filter because the fields count is greater than the
axes count (untagged axis leads to the same bug since always); the splitter
also created less fields than expected and wrote outside of valid memory.

To fixing this bug, the URL splitter must known the correct fields count
instead of deducing it from the tagged axes count.

close !298