[Inendi] fix segfault bombing under layerstack changes
In some circumstances, when changing layers visibility or order, a lots of
crash report were popped up without disturbing the normal use of Inspector.

After analysis, the crashes occur in the PVZoneTree update done in a TBB
parallel_for because the used selection was a wrong pointer. This pointer
invalidation origin was in PVView::process_post_filter_layer(...) which did a
move assignment; it was a copy assignment some time ago.

Doing an inplace operation permits to solve this problem.

So that, this patch introduces in PVCore::PVSelBitFields a template method to
do inplace binary operations and a method named ::inplace_and to implement
the inplace AND op. PVView::process_post_filter_layer(...) uses this later.

close #417