[PVCore] fix deadlocks occuring in PVProgressBox
In some circumstances (and too often on proto-02), cancelling an
investigation loading or stopping'n'processing a source import freeze
entirely Inspector.

In the case of an investigation loading, the deadlock occurs when the main
thread waits for the load thread to end which is waiting for the progressbox
update thread to end but that later tries to synchronize with the main thread
to update the extended label.

To solve those deadlocks, using the asynchonous capability of the Qt
signal/slot system permits to remove the complexity of using mutex/lock.
Instead of using direct connection (which is the default QOject::connect(...)
behavior), the queued mode makes sure the slot is called in the receiver
thread (while direct mode does it in the sender thread). As the processing
function can require user interaction, the ::exec_gui logic uses the blocking
queued mode.

As PVProgressBox headers inclusions has been cleaned-up, other headers files
miss boost/thread.hpp. Those files headers copyright have been updated too.

close #429